ai-ml/chatgpt/agent/html2ppt/requirements.txt
# Core dependencies for HTML2PPT
python-pptx>=0.6.21
beautifulsoup4>=4.11.0
lxml>=4.9.0
Pillow>=9.0.0
requests>=2.28.0
pyyaml>=6.0

# Optional dependencies for advanced features
# Uncomment the following lines if you want AI-powered semantic analysis
# openai>=1.0.0
# transformers>=4.30.0
# torch>=2.0.0

# Utility dependencies
cssutils>=2.6.0
html2text>=2020.1.16
markdown>=3.4.0

# Development dependencies (optional)
# pytest>=7.0.0
# flake8>=5.0.0
# black>=22.0.0